# docs/swagger.yaml
basePath: /
definitions:
...
  domain.ErrorResponse:
    properties:
      errors:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      message:
        type: string
    type: object
...
  domain.RegisterRequest:
    properties:
      email:
        type: string
      full_name:
        minLength: 2
        type: string
      password:
        minLength: 6
        type: string
      phone_number:
        type: string
    required:
    - email
    - full_name
    - password
    - phone_number
    type: object
...
  domain.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      phone_number:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API for Check-In/Out Service
  title: Check-In Service API
  version: "1.0"
paths:
...
  /me:
    get:
      consumes:
      - application/json
      description: Get profile of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Get current user profile
      tags:
      - User
...
  /update-profile:
    put:
      consumes:
      - application/json
      description: Update profile of the authenticated user
      parameters:
      - description: Update Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: invalid request body or validation errors
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Update user profile
      tags:
      - User
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"